move activity to ui package
