+++ code/app/src/androidTest/java/com/example/experiment_automata/BinomialTrialTests.java
@@ -18,6 +18,7 @@ import android.widget.TextView;
 import com.example.experiment_automata.backend.experiments.Experiment;
 import com.example.experiment_automata.backend.experiments.ExperimentMaker;
 import com.example.experiment_automata.backend.experiments.ExperimentType;
+import com.example.experiment_automata.ui.NavigationActivity;
 import com.robotium.solo.Solo;
 
 import org.junit.Before;
