+++ code/app/src/androidTest/java/com/example/experiment_automata/QuestionsUserStoriesTests.java
@@ -16,6 +16,7 @@ package com.example.experiment_automata;
 import android.view.View;
 import android.widget.EditText;
 
+import com.example.experiment_automata.ui.NavigationActivity;
 import com.robotium.solo.Solo;
 
 import org.junit.Before;
