+++ code/app/src/main/java/com/example/experiment_automata/ui/experiments/AddExperimentFragment.java
@@ -21,7 +21,7 @@ import android.widget.Switch;
 import com.example.experiment_automata.backend.experiments.Experiment;
 import com.example.experiment_automata.backend.experiments.ExperimentMaker;
 import com.example.experiment_automata.backend.experiments.ExperimentType;
-import com.example.experiment_automata.NavigationActivity;
+import com.example.experiment_automata.ui.NavigationActivity;
 import com.example.experiment_automata.R;
 import com.example.experiment_automata.backend.users.User;
