+++ code/app/src/main/java/com/example/experiment_automata/ui/home/HomeFragment.java
@@ -16,7 +16,7 @@ import androidx.navigation.Navigation;
 
 import com.example.experiment_automata.backend.experiments.Experiment;
 import com.example.experiment_automata.ui.experiments.NavExperimentDetailsFragment;
-import com.example.experiment_automata.NavigationActivity;
+import com.example.experiment_automata.ui.NavigationActivity;
 import com.example.experiment_automata.R;
 import com.example.experiment_automata.ui.Screen;
 import com.example.experiment_automata.ui.experiments.ExperimentListAdapter;
