+++ code/app/src/main/java/com/example/experiment_automata/ui/trials/add/AddCountTrialFragment.java
@@ -10,7 +10,7 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.example.experiment_automata.backend.experiments.CountExperiment;
-import com.example.experiment_automata.NavigationActivity;
+import com.example.experiment_automata.ui.NavigationActivity;
 import com.example.experiment_automata.R;
 
 /**
